Fix format specifier for profile_hw_io on mips64 (uint32_t is long)

+++ emu.c
@@ -295,9 +295,9 @@ int main(int argc, char *argv[]) {
 		#ifdef N64
 		uint32_t emu_time = TICKS_DISTANCE(t0, TICKS_READ());
 
-		debugf("[PROFILE] cpu:%.2f%% io:%u draw:%.2f%% dma:%.2f%% PC:%06lx\n",
+		debugf("[PROFILE] cpu:%.2f%% io:%lu draw:%.2f%% dma:%.2f%% PC:%06lx\n",
 			(float)emu_time * 100.f / (float)(TICKS_PER_SECOND / 60),
-			profile_hw_io,
+			(unsigned long)profile_hw_io,
 			(float)render_time * 100.f / (float)(TICKS_PER_SECOND / 60),
 			(float)profile_dma_load * 100.f / (float)(TICKS_PER_SECOND / 60),
 			(uint32_t)emu_pc());
